src/main/java/com/airbnb/property/service/UserService.java
package com.airbnb.property.service;

import com.airbnb.property.model.User;
import com.airbnb.property.repository.UserRepository;
import jakarta.persistence.EntityNotFoundException;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.UUID;

@RequiredArgsConstructor
@Service
public class UserService {

    private final UserRepository userRepository;

    public User getUser(UUID userId) {
        return userRepository.findByUuid(userId)
                .orElseThrow(() -> new EntityNotFoundException("User not found with ID: " + userId));
    }
}
